Use otelhttp instead of otelmux
