+++ opentelemetry-tracing/opentelemetry-go/http/Dockerfile
