+++ opentelemetry-tracing/opentelemetry-go/http/docker-compose.yaml
