+++ opentelemetry-tracing/opentelemetry-go/http/server/go.mod
@@ -3,10 +3,9 @@ module server
 go 1.18
 
 require (
-	github.com/gorilla/mux v1.8.0
 	github.com/signalfx/splunk-otel-go/distro v1.3.1
 	github.com/signalfx/splunk-otel-go/instrumentation/net/http/splunkhttp v1.3.1
-	go.opentelemetry.io/contrib/instrumentation/github.com/gorilla/mux/otelmux v0.39.0
+	go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp v0.39.0
 )
 
 require (
@@ -28,6 +27,7 @@ require (
 	go.opentelemetry.io/otel/exporters/otlp/internal/retry v1.13.0 // indirect
 	go.opentelemetry.io/otel/exporters/otlp/otlptrace v1.13.0 // indirect
 	go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc v1.13.0 // indirect
+	go.opentelemetry.io/otel/metric v0.36.0 // indirect
 	go.opentelemetry.io/otel/sdk v1.13.0 // indirect
 	go.opentelemetry.io/otel/trace v1.13.0 // indirect
 	go.opentelemetry.io/proto/otlp v0.19.0 // indirect
